Added more spaces, fixed all pieces
Added more spaces and changed all of the pieces to snap in place on their correct spaces need to fix a bug where if one piece overlaps one space that's not its space then the piece that was supposed to be there stops moving. Next time I also want to finish figuring out how to keep a score count.