Added blue squares learned how to change texture
Learned how to change the texture of sprites, need to figure out how to randomize and change where each piece should go every time the puzzle is completed and reset the board to go to the next level. Might create a score counter if I don't create new levels. Added blue square to apply to pieces to test and fix changing the texture of the sprites on reset.